CactusBot/pyproject.toml
[tool.poetry]
name = "CactusBot"
version = "0.1.0"
description = "CactusBot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"

[nonebot.plugins]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_status", "nonebot_plugin_help"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
